requirements.txt
langchain
langchain-google-genai
langchain-community
streamlit-mic-recorder
gtts